fractional eligibility and don't get stats before epoch

+++ db/migrate/20120521023210_add_page_totals.rb
@@ -2,7 +2,7 @@ class AddPageTotals < ActiveRecord::Migration
   def change
     create_table "page_totals", :force => true do |t|
       t.integer  "page_id"
-      t.integer  "eligible_weeks",             :default => 0
+      t.float  "eligible_weeks",             :default => 0
       t.integer  "pageviews"
       t.integer  "unique_pageviews"
       t.integer  "entrances"
